optimized: separated overview and detailed view from MainActivity added: DetailedViewActivity
