+++ app/src/main/res/layout/activity_detailed_view.xml
@@ -8,7 +8,7 @@
 
     <android.support.constraint.ConstraintLayout
         android:id="@+id/instructor_overview"
-        android:layout_width="match_parent"
+        android:layout_width="368dp"
         android:layout_height="wrap_content"
         app:layout_constraintHorizontal_bias="0.502"
         app:layout_constraintLeft_toRightOf="@+id/lv_instructor_info"
@@ -17,7 +17,7 @@
         tools:layout_constraintLeft_creator="1"
         tools:layout_constraintRight_creator="1"
         tools:layout_constraintTop_creator="1"
-        android:layout_marginTop="64dp">
+        android:layout_marginTop="8dp">
 
         <TextView
             android:id="@+id/tv_instructor_name"
