+++ app/src/main/res/layout/activity_landing.xml
@@ -37,7 +37,7 @@
     <TextView
         android:id="@+id/tv_as_guest"
         android:layout_width="wrap_content"
-        android:layout_height="0dp"
+        android:layout_height="wrap_content"
         android:layout_marginTop="16dp"
         android:text="@string/proceed_as_guest"
         android:textAllCaps="false"
